pyproject.toml
[project]
name = "CalendarApp"
version = "0.1.0"
description = "Create a variety of calendars with your events and duration data"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "arrow>=1.3.0",
    "drawsvg>=2.4.0",
    "openpyxl>=3.1.0",
    "pandas>=2.2.3",
    "python-dateutil>=2.9.0.post0",
    "pillow>=11.0.0",
    "pyyaml>=6.0",
    "fonttools>=4.50.0",
    "holidays>=0.92",
    "graph-layout>=0.1.8",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
